feat: remove root page.tsx and add social group
