--- src/app/feeds/page.tsx
@@ -1,475 +1,0 @@
-import Image from 'next/image';
-import {
-  Award,
-  Users,
-  ChevronDown,
-  MoreHorizontal,
-  ChevronRight,
-  Star,
-  MessageSquare,
-  Building,
-  Eye,
-  Repeat,
-} from 'lucide-react';
-import type { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  title: 'Feed',
-  description: 'Generated by feed',
-};
-
-function FeedPage() {
-  return (
-    <div className="min-h-screen bg-black">
-      <div className="flex min-h-screen justify-between max-w-6xl mx-auto text-white pt-14">
-        {/* Left Sidebar */}
-        <div className="w-64 border-r border-gray-800 flex flex-col">
-          {/* Group Info */}
-          <div className="px-4 py-2">
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-2">
-                <span>[Group name]</span>
-                <div className="w-2 h-2 rounded-full bg-[#fcb300]"></div>
-              </div>
-              <ChevronDown size={16} />
-            </div>
-
-            <div className="mt-4 flex flex-col items-center">
-              <div className="relative">
-                <Image
-                  src="/profile.png?height=80&width=80"
-                  alt="Profile"
-                  width={80}
-                  height={80}
-                  className="rounded-full border-2 border-[#fcb300]"
-                />
-              </div>
-              <div className="mt-2 text-center">
-                <div className="font-medium">Owner · Kim Heyoung</div>
-                <div className="text-xs text-gray-400">Office of Rep</div>
-                <div className="text-xs flex items-center justify-center gap-1">
-                  {/* <div className="w-3 h-2 bg-red-500"></div> */}
-                  <svg
-                    width="16"
-                    height="12"
-                    viewBox="0 0 16 12"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <g clip-path="url(#clip0_2036_4446)">
-                      <rect width="16" height="12" rx="1" fill="white" />
-                      <path
-                        fill-rule="evenodd"
-                        clip-rule="evenodd"
-                        d="M0 0H16V12H0V0Z"
-                        fill="#E31D1C"
-                      />
-                      <path
-                        d="M16 11V12H0V11H16ZM16 10H0V9H16V10ZM16 8H0V7H16V8ZM16 6H0V5H16V6ZM16 4H0V3H16V4ZM16 2H0V1H16V2Z"
-                        fill="#F7FCFF"
-                      />
-                      <rect width="9" height="7" fill="#2E42A5" />
-                      <path
-                        d="M7.81055 1.10059H8.33496L7.92285 1.5166L8.08203 2.17383L7.56934 1.80469L7.14551 2.10059H7.33496L6.92285 2.5166L7.06445 3.10059H7.36328L7.56934 2.56445L7.81055 3.10059H8.33496L7.92285 3.5166L8.08203 4.17383L7.56934 3.80469L7.14551 4.10059H7.33496L6.92285 4.5166L7.06445 5.10059H7.36328L7.56934 4.56445L7.81055 5.10059H8.33496L7.92285 5.5166L8.08203 6.17383L7.56934 5.80469L7.04004 6.17383L7.21875 5.5166L6.74805 5.10059H6.98047L6.56934 4.80469L6.14551 5.10059H6.33496L5.92285 5.5166L6.08203 6.17383L5.56934 5.80469L5.04004 6.17383L5.21875 5.5166L4.74805 5.10059H4.98047L4.56934 4.80469L4.14551 5.10059H4.33496L3.92285 5.5166L4.08203 6.17383L3.56934 5.80469L3.04004 6.17383L3.21875 5.5166L2.74805 5.10059H2.98047L2.56934 4.80469L2.14551 5.10059H2.33496L1.92285 5.5166L2.08203 6.17383L1.56934 5.80469L1.04004 6.17383L1.21875 5.5166L0.748047 5.10059H1.36328L1.56934 4.56445L1.81055 5.10059H2.05957L2.21875 4.5166L1.74805 4.10059H1.98047L1.56934 3.80469L1.04004 4.17383L1.21875 3.5166L0.748047 3.10059H1.36328L1.56934 2.56445L1.81055 3.10059H2.05957L2.21875 2.5166L1.74805 2.10059H1.98047L1.56934 1.80469L1.04004 2.17383L1.21875 1.5166L0.748047 1.10059H1.36328L1.56934 0.564453L1.81055 1.10059H2.33496L1.92285 1.5166L2.06445 2.10059H2.36328L2.56934 1.56445L2.81055 2.10059H3.05957L3.21875 1.5166L2.74805 1.10059H3.36328L3.56934 0.564453L3.81055 1.10059H4.33496L3.92285 1.5166L4.06445 2.10059H4.36328L4.56934 1.56445L4.81055 2.10059H5.05957L5.21875 1.5166L4.74805 1.10059H5.36328L5.56934 0.564453L5.81055 1.10059H6.33496L5.92285 1.5166L6.06445 2.10059H6.36328L6.56934 1.56445L6.81055 2.10059H7.05957L7.21875 1.5166L6.74805 1.10059H7.36328L7.56934 0.564453L7.81055 1.10059ZM5.14551 4.10059H5.33496L4.92285 4.5166L5.06445 5.10059H5.36328L5.56934 4.56445L5.81055 5.10059H6.05957L6.21875 4.5166L5.74805 4.10059H5.98047L5.56934 3.80469L5.14551 4.10059ZM3.14551 4.10059H3.33496L2.92285 4.5166L3.06445 5.10059H3.36328L3.56934 4.56445L3.81055 5.10059H4.05957L4.21875 4.5166L3.74805 4.10059H3.98047L3.56934 3.80469L3.14551 4.10059ZM6.14551 3.10059H6.33496L5.92285 3.5166L6.06445 4.10059H6.36328L6.56934 3.56445L6.81055 4.10059H7.05957L7.21875 3.5166L6.74805 3.10059H6.98047L6.56934 2.80469L6.14551 3.10059ZM2.14551 3.10059H2.33496L1.92285 3.5166L2.06445 4.10059H2.36328L2.56934 3.56445L2.81055 4.10059H3.05957L3.21875 3.5166L2.74805 3.10059H2.98047L2.56934 2.80469L2.14551 3.10059ZM4.14551 3.10059H4.33496L3.92285 3.5166L4.06445 4.10059H4.36328L4.56934 3.56445L4.81055 4.10059H5.05957L5.21875 3.5166L4.74805 3.10059H4.98047L4.56934 2.80469L4.14551 3.10059ZM3.14551 2.10059H3.33496L2.92285 2.5166L3.06445 3.10059H3.36328L3.56934 2.56445L3.81055 3.10059H4.05957L4.21875 2.5166L3.74805 2.10059H3.98047L3.56934 1.80469L3.14551 2.10059ZM5.14551 2.10059H5.33496L4.92285 2.5166L5.06445 3.10059H5.36328L5.56934 2.56445L5.81055 3.10059H6.05957L6.21875 2.5166L5.74805 2.10059H5.98047L5.56934 1.80469L5.14551 2.10059Z"
-                        fill="#F7FCFF"
-                      />
-                    </g>
-                    <defs>
-                      <clipPath id="clip0_2036_4446">
-                        <rect width="16" height="12" rx="1" fill="white" />
-                      </clipPath>
-                    </defs>
-                  </svg>
-                  <span>Oregon Unite Side</span>
-                </div>
-              </div>
-            </div>
-
-            {/* Tier */}
-            <div className="mt-4">
-              <div className="flex justify-between items-center">
-                <span className="text-sm">Tier</span>
-                <div className="flex items-center gap-1">
-                  <span className="text-sm">Diamond</span>
-                  <div className="w-4 h-4 rounded-full bg-[#fcb300] flex items-center justify-center">
-                    <Star size={10} className="text-black" />
-                  </div>
-                </div>
-              </div>
-              <div className="mt-1 h-1 w-full bg-gray-700 rounded-full">
-                <div className="h-full w-full bg-[#fcb300] rounded-full"></div>
-              </div>
-            </div>
-          </div>
-
-          {/* Navigation */}
-          <nav className="mt-4 px-2">
-            <div className="flex items-center gap-3 px-2 py-2 rounded-md hover:bg-gray-800">
-              <div className="w-5 h-5 rounded-full border border-gray-500 flex items-center justify-center">
-                <Users size={12} />
-              </div>
-              <span className="text-sm">Profile</span>
-            </div>
-            <div className="flex items-center gap-3 px-2 py-2 rounded-md hover:bg-gray-800">
-              <div className="w-5 h-5 rounded-full border border-gray-500 flex items-center justify-center">
-                <MessageSquare size={12} />
-              </div>
-              <span className="text-sm">Threads</span>
-            </div>
-            <div className="flex items-center gap-3 px-2 py-2 rounded-md hover:bg-gray-800">
-              <div className="w-5 h-5 rounded-full border border-gray-500 flex items-center justify-center">
-                <Users size={12} />
-              </div>
-              <span className="text-sm">Manage Group</span>
-            </div>
-            <div className="flex items-center gap-3 px-2 py-2 rounded-md hover:bg-gray-800">
-              <div className="w-5 h-5 rounded-full border border-gray-500 flex items-center justify-center">
-                <Users size={12} />
-              </div>
-              <span className="text-sm">Settings</span>
-            </div>
-          </nav>
-
-          {/* Members */}
-          <div className="mt-6 px-4">
-            <h3 className="text-sm font-medium mb-2">Members</h3>
-            {[1, 2, 3, 4, 5].map((i) => (
-              <div key={i} className="flex items-center gap-2 py-2">
-                <Image
-                  src="/profile.png?height=32&width=32"
-                  alt="Member"
-                  width={32}
-                  height={32}
-                  className="rounded-full"
-                />
-                <span className="text-sm">[Nickname]</span>
-              </div>
-            ))}
-            <div className="mt-2 text-xs text-gray-400 flex items-center">
-              <span>View all</span>
-              <ChevronRight size={14} />
-            </div>
-          </div>
-
-          {/* Recent */}
-          <div className="mt-6 px-4">
-            <div className="flex items-center justify-between">
-              <h3 className="text-sm font-medium">Recent</h3>
-              <ChevronDown size={14} />
-            </div>
-            <div className="mt-2 text-xs text-gray-400">
-              <div className="py-1">Crypto/DAO Treasury Transpare...</div>
-              <div className="py-1">Crypto/DAO Act Investor</div>
-              <div className="py-1">Crypto/DAO Welcome to Protec...</div>
-              <div className="mt-1 flex items-center">
-                <span>View all</span>
-                <ChevronRight size={14} />
-              </div>
-            </div>
-          </div>
-
-          {/* Spaces */}
-          <div className="mt-6 px-4">
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-1">
-                <Building size={14} />
-                <h3 className="text-sm font-medium">Spaces</h3>
-              </div>
-              <ChevronDown size={14} />
-            </div>
-            <div className="mt-2 text-xs text-gray-400">
-              <div className="py-1">Crypto/DAO Treasury Transpare...</div>
-              <div className="py-1">Crypto/DAO Act Investor</div>
-              <div className="py-1">Crypto/DAO Welcome to Protec...</div>
-              <div className="mt-1 flex items-center">
-                <span>View all</span>
-                <ChevronRight size={14} />
-              </div>
-            </div>
-          </div>
-
-          {/* Saved */}
-          <div className="mt-6 px-4">
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-1">
-                <MessageSquare size={14} />
-                <h3 className="text-sm font-medium">Saved</h3>
-              </div>
-              <ChevronDown size={14} />
-            </div>
-            <div className="mt-2 text-xs text-gray-400">
-              <div className="py-1">Crypto/DAO Treasury Transpare...</div>
-              <div className="py-1">Crypto/DAO Act Investor</div>
-              <div className="py-1">Crypto/DAO Welcome to Protec...</div>
-              <div className="mt-1 flex items-center">
-                <span>View all</span>
-                <ChevronRight size={14} />
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* Main Content */}
-        <div className="flex-1 flex">
-          {/* Feed */}
-          <div className="flex-1 border-r border-gray-800">
-            {/* Tabs */}
-            <div className="flex border-b border-gray-800">
-              <div className="px-6 py-3 text-sm font-medium">For you</div>
-              <div className="px-6 py-3 text-sm text-gray-400">Following</div>
-            </div>
-
-            {/* Post Input */}
-            <div className="p-4 border-b border-gray-800">
-              <div className="flex gap-3">
-                <Image
-                  src="/profile.png?height=40&width=40"
-                  alt="Profile"
-                  width={40}
-                  height={40}
-                  className="rounded-full"
-                />
-                <input
-                  type="text"
-                  placeholder="Discuss legislation. Drive change."
-                  className="w-full bg-gray-800 rounded-full pl-10 pr-4 py-1.5 text-sm focus:outline-none focus:ring-1 focus:ring-[#fcb300]"
-                />
-                {/* <div className="flex-1 bg-gray-800 rounded-full px-4 py-2 text-gray-400 text-sm">
-                                    Discuss legislation. Drive change.
-                                </div> */}
-              </div>
-            </div>
-
-            {/* Posts */}
-            {[1, 2, 3].map((i) => (
-              <div key={i} className="border-b border-gray-800 p-4">
-                <div className="flex justify-between">
-                  <div className="flex gap-2 items-center">
-                    <div className="uppercase text-xs font-medium bg-gray-800 px-2 py-0.5 rounded">
-                      Crypto
-                    </div>
-                    {i === 1 && (
-                      <div className="uppercase text-xs font-medium bg-[#fcb300] text-black px-2 py-0.5 rounded">
-                        Active
-                      </div>
-                    )}
-                  </div>
-                  <div className="flex items-center gap-2">
-                    <div className="bg-[#fcb300] text-black text-xs px-2 py-0.5 rounded">
-                      New
-                    </div>
-                    <button>
-                      <MessageSquare size={16} />
-                    </button>
-                    <button>
-                      <MoreHorizontal size={16} />
-                    </button>
-                  </div>
-                </div>
-
-                <div className="mt-2">
-                  <div className="flex items-center gap-2 mb-1">
-                    <div className="w-5 h-5 bg-gray-700 rounded"></div>
-                    <div className="text-sm font-medium">Space name</div>
-                    <div className="w-2 h-2 rounded-full bg-[#fcb300]"></div>
-                  </div>
-                  <h2 className="text-lg font-medium">
-                    DAO Treasury Transparency Act & Crypto Investor Protection
-                    Act
-                  </h2>
-                  <p className="text-sm text-gray-400 mt-1">
-                    Explore powerful artworks that amplify voices for equality,
-                    diversity, and justice. This collection brings...
-                    <span className="text-[#fcb300]"> See more</span>
-                  </p>
-                </div>
-
-                <div className="mt-3">
-                  <div className="rounded-md overflow-hidden">
-                    <Image
-                      src="/post-placeholder.jpg?height=300&width=500"
-                      alt="Post image"
-                      width={500}
-                      height={300}
-                      className="w-full object-cover"
-                    />
-                  </div>
-                </div>
-
-                <div className="mt-3 flex justify-between text-sm text-gray-400">
-                  <div className="flex items-center gap-1">
-                    <Eye size={16} />
-                    <span>{i === 1 ? '705' : '212'}</span>
-                  </div>
-                  <div className="flex items-center gap-1">
-                    <Building size={16} />
-                    <span>Space</span>
-                  </div>
-                  <div className="flex items-center gap-1">
-                    <Eye size={16} />
-                    <span>201</span>
-                  </div>
-                  <div className="flex items-center gap-1">
-                    <Star size={16} />
-                    <span>221K</span>
-                  </div>
-                  <div className="flex items-center gap-1">
-                    <Repeat size={16} />
-                    <span>403</span>
-                  </div>
-                </div>
-
-                <div className="mt-2 text-xs text-gray-500 text-right">
-                  1w ago
-                </div>
-              </div>
-            ))}
-          </div>
-
-          {/* Right Sidebar */}
-          <div className="w-80 p-4">
-            {/* Hot Promotion */}
-            <div>
-              <h3 className="font-medium mb-3">Hot Promotion</h3>
-              <div className="bg-gray-800 rounded-lg p-3">
-                <div className="flex items-center gap-2">
-                  <div className="w-10 h-10 rounded bg-blue-500 flex items-center justify-center">
-                    <Award size={16} />
-                  </div>
-                  <div>
-                    <div className="font-medium">Presidential Election</div>
-                  </div>
-                </div>
-                <div className="mt-2 text-xs text-gray-400 flex items-center">
-                  <span>View all</span>
-                  <ChevronRight size={14} />
-                </div>
-              </div>
-            </div>
-
-            {/* News */}
-            <div className="mt-6">
-              <h3 className="font-medium mb-3">News</h3>
-
-              <div className="mb-4">
-                <h4 className="font-medium text-sm">
-                  Ratel' Launches Digital Asset Policy Comparison Feature Ahead
-                  of 2025 Election
-                </h4>
-                <p className="text-xs text-gray-400 mt-1">
-                  Ratel, a blockchain-based legislative social media platform,
-                  has introduced a new feature that allows voters to...
-                </p>
-                <div className="mt-1 text-xs text-gray-400 flex items-center">
-                  <span>View Detail</span>
-                  <ChevronRight size={14} />
-                </div>
-              </div>
-
-              <div className="mb-4">
-                <h4 className="font-medium text-sm">
-                  Legislative Platform 'Ratel' Records Public Opinion on
-                  Election Pledges via Blockchain
-                </h4>
-                <p className="text-xs text-gray-400 mt-1">
-                  Ratel, a blockchain-based social media platform, has launched
-                  a new feature ahead of South Korea's 2025...
-                </p>
-                <div className="mt-1 text-xs text-gray-400 flex items-center">
-                  <span>View Detail</span>
-                  <ChevronRight size={14} />
-                </div>
-              </div>
-
-              <div className="mb-4">
-                <h4 className="font-medium text-sm">
-                  Decentralized Legislative Platform 'RATEL' Leads Crypto
-                  Regulatory Reform in South Korea
-                </h4>
-                <p className="text-xs text-gray-400 mt-1">
-                  RATEL is the world's first decentralized legislative DAO
-                  platform empowering communities to propose and monitor...
-                </p>
-                <div className="mt-1 text-xs text-gray-400 flex items-center">
-                  <span>View Detail</span>
-                  <ChevronRight size={14} />
-                </div>
-              </div>
-            </div>
-
-            {/* Add to your feed */}
-            <div className="mt-6">
-              <h3 className="font-medium mb-3">Add to your feed</h3>
-
-              <div className="mb-3 flex items-center gap-3">
-                <Image
-                  src="/trump.jpg?height=40&width=40"
-                  alt="Donald Trump"
-                  width={40}
-                  height={40}
-                  className="rounded-full"
-                />
-                <div className="flex-1">
-                  <div className="font-medium text-sm">Donald Trump</div>
-                  <div className="text-xs text-gray-400">
-                    President of the US
-                  </div>
-                </div>
-                <button className="text-xs bg-gray-700 rounded-full px-3 py-1">
-                  + Follow
-                </button>
-              </div>
-
-              <div className="mb-3 flex items-center gap-3">
-                <Image
-                  src="/elon.png?height=40&width=40"
-                  alt="Elon Musk"
-                  width={40}
-                  height={40}
-                  className="rounded-full"
-                />
-                <div className="flex-1">
-                  <div className="font-medium text-sm">Elon Musk</div>
-                  <div className="text-xs text-gray-400">
-                    CEO of Tesla and SpaceX
-                  </div>
-                </div>
-                <button className="text-xs bg-gray-700 rounded-full px-3 py-1">
-                  + Follow
-                </button>
-              </div>
-
-              <div className="mb-3 flex items-center gap-3">
-                <Image
-                  src="/jongsook.png?height=40&width=40"
-                  alt="Jongsook Park"
-                  width={40}
-                  height={40}
-                  className="rounded-full"
-                />
-                <div className="flex-1">
-                  <div className="font-medium text-sm">Jongsook Park</div>
-                  <div className="text-xs text-gray-400">
-                    National Assembly of
-                  </div>
-                </div>
-                <button className="text-xs bg-gray-700 rounded-full px-3 py-1">
-                  + Follow
-                </button>
-              </div>
-
-              <div className="mt-2 text-xs text-gray-400 flex items-center">
-                <span>View all</span>
-                <ChevronRight size={14} />
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* Header - Floating on top */}
-      </div>
-    </div>
-  );
-}
-
-export default FeedPage;
