--- src/app/page.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-/* import { useCookie } from '@/lib/contexts/cookie-context'; */
-import MyHome from './_components/MyHome';
-
-export default function Home() {
-  /* let cookie = useCookie(); */
-
-  /* if (cookie?.isLoggedIn) {
-   *   return <MyHome />;
-   * } */
-
-  return <MyHome />;
-}
